backend/pyproject.toml
[project]
name = "Hacktheroom"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "passlib>=1.7.4",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "ruff>=0.14.14",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
    "aiosqlite>=0.22.1",
    "python-jose[cryptography]>=3.5.0",
    "argon2-cffi>=25.1.0",
    "httpx>=0.28.1",
    "dishka>=1.7.2",
    "mypy>=1.19.1",
    "pytest-asyncio>=0.24.0",
    "loguru>=0.7.3",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pillow>=12.1.1",
    "aioboto3>=15.5.0",
    "fastapi-users[sqlalchemy]>=15.0.4",
    "python-multipart>=0.0.22",
]
[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["src/tests"]
[tool.mypy]
python_version = "3.13"
plugins = [
    "pydantic.mypy",
    "sqlalchemy.ext.mypy.plugin"
]

ignore_missing_imports = true
follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
strict_optional = true
disallow_any_generics = true
[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

